+++ src/actor_ref.rs
@@ -1,5 +1,5 @@
 use crate::actor::{ActorTrait, Handler};
-use crate::actor_config::ActorConfig;
+use crate::actor_config::{ActorConfig, RestartPolicy};
 use crate::message::{MessageEnvelope, MessageEnvelopeTrait, MessageTrait};
 use crossbeam_channel::{unbounded, Receiver, Sender, TryRecvError};
 use std::any::Any;
@@ -7,19 +7,19 @@ use std::borrow::BorrowMut;
 use std::ops::DerefMut;
 use std::sync::{Arc, RwLock};
 use std::panic::{UnwindSafe, AssertUnwindSafe, catch_unwind};
-use crate::actor_ref::HandleResult::{MailboxEmpty, Success, ActorPanic};
 
 pub trait ActorRefTrait: Send + Sync {
-    fn handle(&self) -> HandleResult;
+    fn handle(&mut self);
     fn get_config(&self) -> &ActorConfig;
-    fn reset(&mut self);
+    fn handle_panic(&mut self);
+    fn get_current_state(&self) -> ActorState;
 }
 
-#[derive(PartialEq)]
-pub enum HandleResult {
-    MailboxEmpty,
-    Success,
-    ActorPanic,
+#[derive(PartialEq, Clone)]
+pub enum ActorState {
+    Running,
+    Sleeping,
+    Stopped,
 }
 
 #[derive(Clone)]
@@ -27,30 +27,28 @@ pub struct ActorRef<A>
 where
     A: ActorTrait,
 {
-    actor: Arc<RwLock<A>>,
+    actor: A,
     actor_backup: A,
     actor_config: ActorConfig,
     mailbox_in: Sender<MessageEnvelope<A>>,
     mailbox_out: Receiver<MessageEnvelope<A>>,
+    actor_state: ActorState
 }
 
 impl<A> ActorRefTrait for ActorRef<A>
 where
     A: ActorTrait + Clone + UnwindSafe + 'static,
 {
-    fn handle(&self) -> HandleResult {
+    fn handle(&mut self) {
         let mut m = self.mailbox_out.try_recv();
         if m.is_err() {
-            return MailboxEmpty
+            self.actor_state = ActorState::Sleeping;
+            return;
         }
         let mut msg = m.unwrap();
-        let mut a = self.actor.write().unwrap();
-        let mut ac = a.deref_mut();
-        let result = catch_unwind(AssertUnwindSafe(|| msg.handle(ac)));
-        //msg.handle(ac);
-        match result {
-            Err(err) => ActorPanic,
-            _ => Success
+        let result = catch_unwind(AssertUnwindSafe(|| msg.handle(&mut self.actor)));
+        if result.is_err() {
+            self.handle_panic();
         }
     }
 
@@ -58,8 +56,16 @@ where
         &self.actor_config
     }
 
-    fn reset(&mut self) {
-        self.actor = Arc::new(RwLock::new(self.actor_backup.clone()));
+    fn handle_panic(&mut self) {
+        if self.actor_config.restart_policy == RestartPolicy::Never {
+            self.actor_state = ActorState::Stopped;
+            return;
+        }
+        self.actor = self.actor_backup.clone();
+    }
+
+    fn get_current_state(&self) -> ActorState {
+        self.actor_state.clone()
     }
 }
 
@@ -75,9 +81,10 @@ where
     ) -> Self {
         let actor_backup = actor.clone();
         Self {
-            actor: Arc::new(RwLock::new(actor)),
+            actor,
             actor_backup,
             actor_config,
+            actor_state: ActorState::Running,
             mailbox_in: sender,
             mailbox_out: receiver,
         }
@@ -87,6 +94,8 @@ where
         A: Handler<M>,
         M: MessageTrait + Clone + 'static,
     {
-        self.mailbox_in.send(MessageEnvelope::new(msg));
+        if self.actor_state != ActorState::Stopped {
+            self.mailbox_in.send(MessageEnvelope::new(msg));
+        }
     }
 }